Extracted Extensions  to ObservationManagerMenuExtensions
