+++ observation-manager/src/main/java/de/lehmannet/om/ui/navigation/ObservationManager.java
@@ -7,7 +7,6 @@
 
 package de.lehmannet.om.ui.navigation;
 
-import java.awt.Cursor;
 import java.awt.Dimension;
 import java.awt.Frame;
 import java.awt.Toolkit;
@@ -16,12 +15,10 @@ import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
 import java.awt.event.WindowEvent;
 import java.io.File;
-import java.io.IOException;
 import java.io.PrintStream;
 import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -30,15 +27,12 @@ import java.util.Map;
 import java.util.MissingResourceException;
 import java.util.PropertyResourceBundle;
 import java.util.ResourceBundle;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
 
 import javax.swing.AbstractAction;
 import javax.swing.ImageIcon;
 import javax.swing.JCheckBoxMenuItem;
 import javax.swing.JComponent;
 import javax.swing.JDialog;
-import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JMenu;
 import javax.swing.JMenuBar;
@@ -46,7 +40,6 @@ import javax.swing.JMenuItem;
 import javax.swing.JOptionPane;
 import javax.swing.JSplitPane;
 import javax.swing.KeyStroke;
-import javax.swing.filechooser.FileFilter;
 import javax.swing.plaf.ColorUIResource;
 import javax.swing.plaf.metal.DefaultMetalTheme;
 
@@ -64,9 +57,7 @@ import de.lehmannet.om.IScope;
 import de.lehmannet.om.ISession;
 import de.lehmannet.om.ISite;
 import de.lehmannet.om.ITarget;
-import de.lehmannet.om.ui.dialog.AboutDialog;
 import de.lehmannet.om.ui.dialog.AbstractDialog;
-import de.lehmannet.om.ui.dialog.ExtensionInfoDialog;
 import de.lehmannet.om.ui.dialog.EyepieceDialog;
 import de.lehmannet.om.ui.dialog.FilterDialog;
 import de.lehmannet.om.ui.dialog.LensDialog;
@@ -204,6 +195,7 @@ public class ObservationManager extends JFrame implements ActionListener {
     private final ObservationManagerMenuData menuData;
     private final ObservationManagerMenuExtras menuExtras;
     private final ObservationManagerMenuHelp menuHelp;
+    private final ObservationManagerMenuExtensions menuExtensions;
 
     private File schemaPath;
 
@@ -262,6 +254,7 @@ public class ObservationManager extends JFrame implements ActionListener {
         this.menuData = new ObservationManagerMenuData(this.configuration, this.xmlCache, this);
         this.menuExtras = new ObservationManagerMenuExtras(this.configuration, this.xmlCache, this);
         this.menuHelp = new ObservationManagerMenuHelp(this.configuration, this.xmlCache, this);
+        this.menuExtensions = new ObservationManagerMenuExtensions(this.configuration, this.xmlCache, this);
 
         boolean nightVisionOnStartup = Boolean
                 .parseBoolean(this.configuration.getConfig(ObservationManager.CONFIG_NIGHTVISION_ENABLED, "false"));
@@ -422,12 +415,12 @@ public class ObservationManager extends JFrame implements ActionListener {
                 this.menuExtras.showLogDialog();
             } else if (source.equals(this.updateMenuEntry)) {
                this.menuExtras.checkUpdates();
+            }  else if (source.equals(this.extensionInfo)) {
+                this.menuExtensions.showExtensionInfo();
+            } else if (source.equals(this.installExtension)) {
+                this.menuExtensions.installExtension(null);
             } else if (source.equals(this.aboutInfo)) {
                 this.menuHelp.showInfo();
-            } else if (source.equals(this.extensionInfo)) {
-                this.showExtensionInfo();
-            } else if (source.equals(this.installExtension)) {
-                this.installExtension(null);
             }
         }
 
@@ -463,15 +456,7 @@ public class ObservationManager extends JFrame implements ActionListener {
 
     
     
-    private void showExtensionInfo() {
-
-        if (this.extLoader.getExtensions().isEmpty()) {
-            this.createInfo(ObservationManager.bundle.getString("info.noExtensionsInstalled"));
-        } else {
-            new ExtensionInfoDialog(this);
-        }
-
-    }
+    
 
     public void reloadLanguage() {
 
@@ -708,6 +693,10 @@ public class ObservationManager extends JFrame implements ActionListener {
             }
         }
 
+    }
+
+    public void exit() {
+        this.menuFile.exit(this.changed);
     }
 
     public void updateLeft() {
@@ -959,49 +948,7 @@ public class ObservationManager extends JFrame implements ActionListener {
      */
     // }
 
-    private boolean checkWriteAccess(ZipFile zipFile, File destinationRoot) {
 
-        Enumeration enumeration = zipFile.entries();
-        ZipEntry ze;
-
-        // Unpack all the ZIP file entries into install dir
-        File currentFile;
-        boolean result = true;
-        while (enumeration.hasMoreElements()) {
-
-            ze = (ZipEntry) enumeration.nextElement();
-            currentFile = this.getDestinationFile(ze.getName(), destinationRoot, false);
-
-            if (currentFile != null) {
-                while (!currentFile.exists()) { // New file/folder, which
-                                                // doesn't exist so far
-                    currentFile = new File(currentFile.getParent()); // Check
-                                                                     // write
-                                                                     // permission
-                                                                     // on
-                                                                     // parent
-                }
-
-                if (!currentFile.canWrite()) { // We've found at least one file,
-                                               // which we would need to
-                                               // overwrite, but do not
-                                               // have the permission to
-                    System.err.println("Write check failed for: " + currentFile);
-                    result = false;
-                }
-            }
-
-        }
-
-        return result;
-
-    }
-
-    public boolean checkWriteAccess(File file) {
-
-        return file.canWrite();
-
-    }
 
     // ---------------
     // Private Methods --------------------------------------------------------
@@ -1053,21 +1000,7 @@ public class ObservationManager extends JFrame implements ActionListener {
      * }
      */
 
-    private File getDestinationFile(String filename, File destinationFolder, boolean removeRootFolder) {
-
-        if (removeRootFolder) {
-            // Remove root folder
-            filename = filename.substring(filename.indexOf("/") + 1);
-
-            if ("".equals(filename)) { // That must have been the root folder
-                return null;
-            }
-        }
-
-        return new File(
-                destinationFolder.getAbsolutePath() + File.separator + /* "testing" + File.separator + */filename);
-
-    }
+    
 
     // ****************************************************************
     // Only required for Auto. Update which is currently not supported
@@ -1592,125 +1525,7 @@ public class ObservationManager extends JFrame implements ActionListener {
 
     }
 
-    private void installExtension(File[] files) {
-
-        // No files passed, so need to ask user for list of extensions
-        if (files == null) {
-
-            // Let user choose extension zip file
-            JFileChooser chooser = new JFileChooser(ObservationManager.bundle.getString("extenstion.chooser.title"));
-            FileFilter zipFileFilter = new FileFilter() {
-                @Override
-                public boolean accept(File f) {
-                    return (f.getName().endsWith(".ome")) || (f.isDirectory());
-                }
-
-                @Override
-                public String getDescription() {
-                    return "Observation Manager extensions";
-                }
-            };
-            chooser.setFileFilter(zipFileFilter);
-            String last = this.configuration.getConfig(ObservationManager.CONFIG_LASTDIR);
-            if ((last != null) && !("".equals(last.trim()))) {
-                File dir = new File(last);
-                if (dir.exists()) {
-                    chooser.setCurrentDirectory(dir);
-                }
-            }
-            chooser.setMultiSelectionEnabled(true);
-            int returnVal = chooser.showOpenDialog(this);
-            if (returnVal == JFileChooser.APPROVE_OPTION) {
-                files = chooser.getSelectedFiles();
-            } else {
-                return;
-            }
-
-        }
-
-        // Check whether deployment can be done -> whether we've write
-        // permissions for all files
-        StringBuilder negativeResult = new StringBuilder();
-        List<File> filesOK = new ArrayList<>();
-        try {
-            boolean checkResult = false;
-            for (File file : files) {
-                checkResult = this.checkWriteAccess(new ZipFile(file), this.installDir.getInstallDir());
-                if (!checkResult) {
-                    negativeResult.append(" ").append(file.getName());
-                } else {
-                    filesOK.add(file);
-                }
-            }
-        } catch (IOException ioe) {
-            System.out.println("Error while checking extension zip file. Zip file may be corrupted.\n" + ioe);
-        }
-        File[] filesCheckedOK = (File[]) filesOK.toArray(new File[] {});
-
-        // --- Start with deployment
-
-        Cursor hourglassCursor = new Cursor(Cursor.WAIT_CURSOR);
-        setCursor(hourglassCursor);
-
-        StringBuilder positiveResult = new StringBuilder();
-        int successCounter = 0;
-        for (int i = 0; i < filesCheckedOK.length; i++) {
-            try {
-                positiveResult.append(" ").append(this.extLoader.addExtension(new ZipFile(filesCheckedOK[i])));
-                successCounter++;
-                if (i < filesCheckedOK.length - 1) { // There is at least one
-                                                     // more ZIP to add
-                    positiveResult.append(", ");
-                }
-            } catch (IOException ioe) {
-                System.out.println("Error in extension zip file. Zip file may be corrupted.\n" + ioe);
-
-                Cursor normalCursor = new Cursor(Cursor.DEFAULT_CURSOR);
-                setCursor(normalCursor);
-
-                negativeResult.append(" ").append(filesCheckedOK[i].getName());
-            }
-        }
-
-        boolean result = false;
-
-        // Show all positive results
-        if (successCounter > 0) {
-            this.createInfo(ObservationManager.bundle.getString("info.addExtensionSuccess") + " " + positiveResult);
-
-            // Until we found a better way to handle extension, we need to
-            // restart... :-(
-            if (true) {
-                this.createInfo(ObservationManager.bundle.getString("info.addExtensionRestart"));
-                // this.exit();
-                this.menuFile.exit(this.changed);
-            }
-
-            result = true;
-        }
-
-        // Show all negative results
-        if (successCounter < files.length) { // We check here against the
-                                             // original files Array, to see
-                                             // whether we had some
-                                             // problems during check OR
-                                             // installation
-            this.createWarning(ObservationManager.bundle.getString("error.addExtensionFail") + " " + negativeResult);
-
-            result = false;
-        }
-
-        // Inform about restart (if any installation was successfull)
-        /*
-         * if( successCounter > 0 ) { // Until we found a better way to handle
-         * extension, we need to restart... :-( this.createInfo(ObservationManager
-         * .bundle.getString("info.addExtensionRestart")); this.exit(); }
-         */
-
-        Cursor normalCursor = new Cursor(Cursor.DEFAULT_CURSOR);
-        setCursor(normalCursor);
-
-    }
+
